fix(foundry-api-bridge): centralize randomUUID fallback, also wrap try/catch
Two call sites needed the secure-context fallback (WebSocketClient and
dialog-intercept). Extract to transport/random-id.ts so we can't miss the
next one. Also wrap crypto.randomUUID() in try/catch — some browsers
expose the function in non-secure contexts but throw when it's called.
