+++ apps/foundry-api-bridge/src/transport/WebSocketClient.ts
@@ -1,4 +1,5 @@
 import type { Command, CommandResponse } from '@/commands';
+import { generateRandomId } from './random-id';
 
 export interface WebSocketClientConfig {
   url: string;
@@ -249,17 +250,5 @@ export class WebSocketClient {
 }
 
 function generateBridgeId(): string {
-  // crypto.randomUUID requires a secure context (HTTPS or localhost). When
-  // Foundry is served over plain HTTP from a non-localhost host (e.g.
-  // http://server.ad:30000) the Web Crypto API is unavailable and the call
-  // throws. Fall back to a UUID v4 generated from Math.random — collision
-  // probability is irrelevant for short-lived bridge correlation IDs.
-  if (typeof crypto !== 'undefined' && typeof crypto.randomUUID === 'function') {
-    return crypto.randomUUID();
-  }
-  return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, (c) => {
-    const r = (Math.random() * 16) | 0;
-    const v = c === 'x' ? r : (r & 0x3) | 0x8;
-    return v.toString(16);
-  });
+  return generateRandomId();
 }
